Fix unsignedInt and positiveInt representations
The r4 spec for FHIR incorrectly states the underlying type of
unsignedInt and positiveInt to be strings, despite the representation
in JSON being a number. The Fhenix tool needed to be updated to
address this.

This commit updates the unsignedInt and positiveInt types to be
represented as numbers in the generated Go code.